* master:
  Summary: Finish adding in titles for fragments Test Plan:
  Summary: Adding titles to fragments Test Plan:
  Summary: Changing schools so they request the school on refresh Test Plan:
  changed profile text to director if director and removed inactive users from browse mentors page
  Summary: Fixed bug where usr list wouldn't show up
  Summary: fixing a bug where mentors don't show up
  <one-line short summary, no longer than this text>
  slightly better
  cleaned up
  thoroughly tested
  added empty abstract activity skeleton getting closer to finishing school parallax view
  remove location
  added empty abstract activity skeleton

+++ sage-android/Sage/app/src/main/java/blueprint/com/sage/utility/view/ViewUtils.java
@@ -34,7 +34,7 @@ public class ViewUtils {
                                                        .getColor(color, activity.getTheme()));
     }
 
-    public static void setToolBarElevation(Activity activity, int elevation) {
+    public static void setToolBarElevation(Activity activity, float elevation) {
         ActionBar actionBar = ((AppCompatActivity) activity).getSupportActionBar();
         if (actionBar != null)
             actionBar.setElevation(elevation);
